+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Percollate Changelog
 
+# 0.1.5
+
+-   Only replace single image SRCs with their parent's HREF if the latter matches an image file
+
 # 0.1.3
 
 -   Introduces separate commands. For now, the `pdf` command is implemented.
@@ -7,4 +11,4 @@
 
 # 0.1.2
 
-Run `percollate` as a CLI tool that takes one or more URLs and bundles them into a single PDF file.
+-   Run `percollate` as a CLI tool that takes one or more URLs and bundles them into a single PDF file.